backend/requirements.txt
# Core dependencies
fastapi>=0.104.1,<0.110.0
uvicorn[standard]>=0.25.0,<0.30.0

# Database
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0  # PostgreSQL driver

# Pydantic for data validation
pydantic>=2.5.2,<3.0.0

# PDF/DOCX parsing
PyPDF2>=3.0.1,<4.0.0
pdfplumber>=0.8.0,<0.9.0
python-docx>=1.0.0,<2.0.0

# File handling
beautifulsoup4>=4.12.3,<5.0.0

# HTTP requests
httpx>=0.26.0,<0.30.0

# Ollama integration
ollama>=0.5.1,<0.6.0
langchain-ollama>=0.2.17,<0.3.0

# Testing
pytest>=8.0.0,<9.0.0
httpx>=0.26.0,<0.30.0  # For testing

# Development tools
ruff>=0.3.2,<0.4.0
black>=23.12.1,<25.0.0

# Type checking
mypy>=1.10.0,<2.0.0

# Environment management
python-dotenv>=1.0.0,<2.0.0
